+++ exporters/otlp/otlptrace/otlptracegrpc/internal/x/x.go
@@ -5,25 +5,12 @@
 //
 // This package should only be used for features defined in the specification.
 // It should not be used for experiments or new project ideas.
-package x // import "go.opentelemetry.io/otel/sdk/internal/x"
+package x // import "go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc/internal/x"
 
 import (
 	"os"
 	"strings"
 )
-
-// Resource is an experimental feature flag that defines if resource detectors
-// should be included experimental semantic conventions.
-//
-// To enable this feature set the OTEL_GO_X_RESOURCE environment variable
-// to the case-insensitive string value of "true" (i.e. "True" and "TRUE"
-// will also enable this).
-var Resource = newFeature("RESOURCE", func(v string) (string, bool) {
-	if strings.ToLower(v) == "true" {
-		return v, true
-	}
-	return "", false
-})
 
 // SelfObservability is an experimental feature flag that determines if SDK
 // self-observability metrics are enabled.