+++ sdk/trace/tracer.go
@@ -100,8 +100,10 @@ func (tr *tracer) newSpan(ctx context.Context, name string, config *trace.SpanCo
 	}
 	if isSampled(samplingResult) {
 		scc.TraceFlags = psc.TraceFlags() | trace.FlagsSampled
+		tr.provider.spanLiveCount.Add(ctx, 1, tr.provider.sampledAttributes)
 	} else {
 		scc.TraceFlags = psc.TraceFlags() &^ trace.FlagsSampled
+		tr.provider.spanLiveCount.Add(ctx, 1, tr.provider.notSampledAttributes)
 	}
 	sc := trace.NewSpanContext(scc)